requirements.txt
Django==3.2
mysqlclient==1.4.6
django-tastypie==0.14.3
lxml==4.6.3
defusedxml==0.6.0
jsonfield==3.1.0
django-elasticsearch-dsl==7.2.1
elasticsearch==7.15.1
elasticsearch-dsl==7.4.0
pyparsing==3.0.9
six==1.16.0
#deform y colander se pueden quitar si se comenta su import en thesaurus/field_definitions*.py
deform==2.0.15
colander==1.8.3
#python-dateutil==2.8.2
#python-mimeparse==1.6.0
#pytz==2021.3
#django-multiselectfield==0.1.12
#django-crum==0.7.8
#simplejson==3.17.0
#requests==2.24.0
#sqlparse==0.4.2
#urllib3==1.25.2



